+++ sdk/servicebus/service-bus/src/serializers/topicResourceSerializer.ts
@@ -107,6 +107,8 @@ export function buildTopicRuntimeProperties(rawTopic: Record<string, any>): Topi
     name: getString(rawTopic[Constants.TOPIC_NAME], "topicName"),
     sizeInBytes: getIntegerOrUndefined(rawTopic[Constants.SIZE_IN_BYTES]),
     subscriptionCount: getIntegerOrUndefined(rawTopic[Constants.SUBSCRIPTION_COUNT]),
+    sqlFilterCount: getIntegerOrUndefined(rawTopic[Constants.SQL_FILTER_COUNT]),
+    correlationFilterCount: getIntegerOrUndefined(rawTopic[Constants.CORRELATION_FILTER_COUNT]),
     createdAt: getDate(rawTopic[Constants.CREATED_AT], "createdAt"),
     scheduledMessageCount: getMessageCountDetails(rawTopic[Constants.COUNT_DETAILS])
       .scheduledMessageCount,
@@ -447,6 +449,16 @@ export interface TopicRuntimeProperties {
    *
    */
   subscriptionCount?: number;
+
+  /**
+   * The total number of SQL filters across all of the topic's subscriptions.
+   */
+  sqlFilterCount?: number;
+
+  /**
+   * The total number of correlation filters across all of the topic's subscriptions.
+   */
+  correlationFilterCount?: number;
 
   /**
    * The number of scheduled messages.