+++ sdk/servicebus/service-bus/src/serviceBusAtomManagementClient.ts
@@ -125,11 +125,11 @@ export interface ServiceBusAdministrationClientOptions extends CommonClientOptio
   /**
    * Service version of the ATOM API.
    *
-   * Currently supported = "2021-05" | "2017-04"
+   * Currently supported = "2024-05" | "2021-05" | "2017-04"
    *
-   * Defaults to "2021-05".
+   * Defaults to "2024-05".
    */
-  serviceVersion?: "2021-05" | "2017-04";
+  serviceVersion?: "2024-05" | "2021-05" | "2017-04";
 }
 
 /**